added some fun clipart, restructured some data layer code into model

+++ app/Http/Controllers/BooksController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Books;
 use App\Booklist;
+use Session;
 use Request;
 
 class BooksController extends Controller
@@ -27,34 +28,19 @@ class BooksController extends Controller
     }
 
     public function finalize() {
-        $this->data['booklists'] = Booklist::all();
-        $books = Books::all();
-        $this->data['books_count'] = count($books);
+        $this->data['booklists'] = Booklist::getforweb();
+        $this->data['reading_queue_listid'] = Booklist::reading_queue_listid();
+        $this->data['reading_queue_books_count'] = Booklist::reading_queue_books_count();
+        $this->data['reading_queue_books_pages_count'] = Booklist::reading_queue_books_pages_count();
+        $this->data['books_count'] = Books::count();
+        $this->data['pages_count'] =  Books::sum('page_count');
         $this->data['booklists_count'] = count($this->data['booklists']);
+
+        if($this->data['books_count'] == 0) {
+            Session::flash('message', 'There are no books created, add your favorites today!');
+        }        
     }
 
-	public function google_api_search($query) {
-        try {
-            $client = new \Google_Client();
-            $client->setApplicationName(env("APP_NAME"));
-            $client->setDeveloperKey(env("GOOGLE_API_KEY"));
-
-            $service = new \Google_Service_Books($client);
-            $optParams = array();   
-            $google_results = $service->volumes->listVolumes($query, $optParams);
-
-            $res = [];
-
-            foreach ($google_results as $item) {
-                $res[] = $item['volumeInfo'];
-            }
-
-        } catch (Exception $e) {
-            
-        }
-        
-        return $res;
-	}
 
     public function start() {
 
@@ -63,6 +49,22 @@ class BooksController extends Controller
         $this->finalize();
         return view($this->data['view'], $this->data );
     }       
+
+    public function bookplanner() {
+
+        $form_data = Request::all();   
+        $this->data['pagetitle'] = 'boojbooks | My Reading Planner.';    
+        $this->data['view'] = 'home';
+        $this->data['nav_class'] = "planner";
+        // 99999 == Reading Queue, cheap hardcode :/
+
+        $booklist = Booklist::reading_queue_list();
+        $books = Books::getbybooklist($booklist->id);
+        $results = $books;  
+        $this->data['search_results'] = $results;                  
+        $this->finalize();
+        return view($this->data['view'], $this->data );
+    }     
 
     public function about() {
 
@@ -87,81 +89,16 @@ class BooksController extends Controller
 
         if(isset($form_data['query_api']) && !empty($form_data['query_api'])) {
     		if(isset($form_data['query']) && !empty($form_data['query'])) {
-    			$results = $this->google_api_search($form_data['query']);
-                foreach($results as &$item) {
-                    $item['book_id'] = -1;
-
-                    if(isset($item['industryIdentifiers'][0])) {
-                        $item['book_id'] = $item['industryIdentifiers'][0]['identifier'];
-                    }
-
-                    $book = Books::where('external_id', '=', $item['book_id'])->get();
-
-                    if(count($book)==0) { 
-                        $book = new Books;
-                        $book->booklist_id = 0;
-                        $book->notes = "";
-                        $book->pagenum = "";
-                    } else {
-                        $book = Books::find($book[0]['id']);
-                    }
-
-                    $item['booklist_id'] = $book->booklist_id;
-
-                    $t = "";
-                    if(is_array($item['authors'])) {
-                        foreach ($item['authors'] as $author) {
-                            $t .= ", " . $author;
-                        }
-                    } 
-
-                    if($t) { 
-                        $book->author = $t;
-                    }
-
-                    if(substr($book->author, 0, 2) == ', ') {
-                        $book->author = substr($book->author, 2);
-                    }
-
-                    $t = "";
-                    if(is_array($item['categories'])) {
-                        foreach ($item['categories'] as $category) {
-                            $t .= ", " . $category;
-                        }
-                    } 
-
-                    if($t) { 
-                        $book->category = $t;
-                    }                
-
-                    if(substr($book->category, 0, 2) == ', ') {
-                        $book->category = substr($book->category, 2);
-                    }
-
-                    $item->author = $book->author;
-                    $book->external_id = $item['book_id'];
-                    $book->page_count = ($item['pageCount'] ? $item['pageCount'] : 0);
-                    $book->description = ($item['description'] ? $item['description'] : "");
-                    $book->name = $item['title'];
-                    $book->imageurl = ($item['imageLinks']['thumbnail'] ? $item['imageLinks']['thumbnail'] : "");
-                    $book->infourl = $item['infoLink'];
-                    $book->category = ($book['category'] ? $book['category']  : "");
-                    $book->author = ($book['author'] ? $book['author']  : "");
-                    $item['external_id'] = $item['book_id'];
-                    $item['book_id'] = $book['id'];
-                    $book->save();             
-                 }
+    			Books::google_api_search($form_data['query']);
     		} else { 
                 $form_data['query'] = "";
             }
-        } else {
-            $books = Books::where('name', 'like', '%' . $form_data['query'] . '%')
-                ->orderBy('name', 'asc')
-                ->get();
-            $results = $books->toArray();
         }
 
-        $booklists = Booklist::all();
+        $books = Books::getforweb();
+        $results = $books;
+
+        $booklists = Booklist::getforweb();
         if($this->data['view'] == 'lists'){
             $results = $booklists;
         }
@@ -178,24 +115,44 @@ class BooksController extends Controller
         $this->data['view'] = 'books';
 
         $book = Books::find($book_id);
-        $booklists = Booklist::all();
+        $booklists = Booklist::getforweb();
 
         if(isset($form_data['save_book']) && !empty($form_data['save_book'])) {
             if(isset($form_data['delete']) && !empty($form_data['delete'])) {
                 Books::destroy($book_id);
                 $this->finalize();
+                Session::flash('message', 'Book successfully deleted.');
                 return redirect($this->base_url);
             } else {
                 if(!$book || $book_id == 0) {
-                    $book = new Books();
-                    $book->external_id = "0";
+                    if(isset($form_data['new']) && !empty($form_data['new'])) {
+                        $book = new Books();
+                        $book->external_id = "0";
+                        $book->booklist_id = -1;
+                        $book->name = "";
+                        $book->author = "";
+                        $book->description = "";
+                        $book->notes = "";
+                        $book->page_num = 0;
+                        $book->page_count = 0;
+                        $book->time_to_read = 0;
+                        $book->infourl = "";
+                        $book->imageurl = "";
+                        $book->category = "";
+                        $s_message = "New book successfully saved.";
+                    }
                 } else {
+                    $s_message = "Book successfully saved.";
                     $book = Books::find($book_id);
                 }
             }           
 
-            if(isset($form_data['booklist_id']) && !empty($form_data['booklist_id'])) {
-                $book->booklist_id = $form_data['booklist_id'];
+            if(isset($form_data['add_to_queue']) && !empty($form_data['add_to_queue'])) {
+                $book->booklist_id = Booklist::reading_queue_listid();
+            } else {
+                if(isset($form_data['booklist_id']) && !empty($form_data['booklist_id'])) {
+                    $book->booklist_id = $form_data['booklist_id'];
+                }
             }
             if(isset($form_data['name']) && !empty($form_data['name'])) {
                 $book->name = $form_data['name'];
@@ -209,10 +166,11 @@ class BooksController extends Controller
             if(isset($form_data['notes']) && !empty($form_data['notes'])) {
                 $book->notes = $form_data['notes'];
             } 
-            if(isset($form_data['pagenum']) && !empty($form_data['pagenum'])) {
-                $book->pagenum = $form_data['pagenum'];
+            if(isset($form_data['page_num']) && !empty($form_data['page_num'])) {
+                $book->page_num = $form_data['page_num'];
             }                                 
             $book->save();
+            Session::flash('message', $s_message);
         }
 
         $this->data['book'] = $book;
@@ -230,9 +188,13 @@ class BooksController extends Controller
             $form_data['booklist_id'] = 0;
         }
 
-        $books = Books::where('booklist_id', '=', $booklist_id)->get();
+        $books = Books::getbybooklist($booklist_id);
 
-        $this->data['search_results'] = $books->toArray();
+        if(count($books) === 0) {
+            Session::flash('message', 'Sorry Chap, no books here!  Search for and add some now!');
+        }
+
+        $this->data['search_results'] = $books;
         $this->data['nav_class'] = "lists";
         $this->finalize();
         return view($this->data['view'], $this->data );
@@ -244,38 +206,45 @@ class BooksController extends Controller
         $this->data['view'] = 'lists';
 
         if(!isset($form_data['booklist_id']) || empty($form_data['booklist_id'])) {
-            $form_data['booklist_id'] = 0;
+            $form_data['booklist_id'] = -1;
         }
 
         if(isset($form_data['delete']) && !empty($form_data['delete'])) {
             $this->data['view'] = 'lists';
             $booklist = Booklist::find($booklist_id);
             if($booklist) {
-                $booklist->delete();
+                if($booklist->listname != "- Reading Queue -") {
+                    $booklist->delete();
+                    Session::flash('message', 'Booklist successfully deleted.');
+                }
+                return redirect('booklists');
             }
         }
 
         if(isset($form_data['newlist']) && !empty($form_data['newlist'])) {
             $this->data['view'] = 'lists';
-            $booklist = new Booklist;
-            $booklist->listname = $form_data['newlist'];
-            $booklist->save();
+            if($form_data['newlist'] != "- Reading Queue -") {
+                $booklist = new Booklist;
+                $booklist->listname = $form_data['newlist'];
+                $booklist->save();
+                Session::flash('message', 'New Booklist successfully saved.');
+            }
         }        
 
         if(isset($form_data['mylists']) && !empty($form_data['mylists'])) {
             $this->data['view'] = 'lists';
         }        
 
-        $booklists = Booklist::all();
-        $booklist_script = "<select name='booklist_id'>";
-        foreach ( $booklists as $booklist ) {
-            $booklist_script .= "<option ". ($form_data['booklist_id'] == $booklist->id ? 'SELECTED' : '') . " value='" . $booklist->id . "'>" . $booklist->listname ."</option>";
-        }
-        $booklist_script .= "</select>";
+        $booklists = Booklist::getforweb();
 
         if($this->data['view'] == 'lists'){
             $results = $booklists;
         }
+
+        if(count($booklists) == 0) {
+            Session::flash('message', 'There are no booklists created, create one below!');
+        }
+
 
         $this->data['search_results'] = $results;
         $this->data['nav_class'] = "lists";
